Fix warning in Script.cpp: Change variable to size_t.

+++ src/scripting/Script.cpp
@@ -1612,7 +1612,7 @@ long GetSystemVar(EERIE_SCRIPT * es,INTERACTIVE_OBJ * io, const std::string& _na
 
 			if (!specialstrcmp(name, "^PLAYERCASTING"))
 			{
-				for (long i = 0; i < MAX_SPELLS; i++)
+				for (size_t i = 0; i < MAX_SPELLS; i++)
 				{
 					if (spells[i].exist)
 					{
